requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.4
packaging==24.1
sqlparse==0.5.1
tzdata==2024.2
